Rename _overlap -> overlap
